+++ src/binxml/tokens.rs
@@ -17,11 +17,13 @@ use std::io::Seek;
 use std::io::SeekFrom;
 
 use crate::evtx_chunk::EvtxChunk;
+use encoding::EncodingRef;
 use std::borrow::Cow;
 
 pub fn read_template<'a>(
     cursor: &mut Cursor<&'a [u8]>,
     chunk: Option<&'a EvtxChunk<'a>>,
+    ansi_codec: EncodingRef,
 ) -> Result<BinXmlTemplate<'a>> {
     trace!("TemplateInstance at {}", cursor.position());
 
@@ -57,13 +59,13 @@ pub fn read_template<'a>(
 
         cursor.seek(SeekFrom::Start(u64::from(template_definition_data_offset)))?;
 
-        let template_def = read_template_definition(cursor, chunk)?;
+        let template_def = read_template_definition(cursor, chunk, ansi_codec)?;
 
         cursor.seek(SeekFrom::Start(position_before_seek))?;
 
         Cow::Owned(template_def)
     } else {
-        Cow::Owned(read_template_definition(cursor, chunk)?)
+        Cow::Owned(read_template_definition(cursor, chunk, ansi_codec)?)
     };
 
     let number_of_substitutions = try_read!(cursor, u32);
@@ -97,11 +99,12 @@ pub fn read_template<'a>(
             descriptor.value_type,
             position_before_reading_value
         );
-        let value = BinXmlValue::deserialized_sized_value_type(
+        let value = BinXmlValue::deserialize_value_type(
             &descriptor.value_type,
             cursor,
             chunk,
-            descriptor.size,
+            Some(descriptor.size),
+            ansi_codec,
         )?;
 
         trace!("\t {:?}", value);
@@ -123,7 +126,7 @@ pub fn read_template<'a>(
                   expected_position,
                   &descriptor);
 
-            cursor.seek(SeekFrom::Start((current_position + u64::from(diff)) as u64))?;
+            cursor.seek(SeekFrom::Start((current_position + diff) as u64))?;
         }
         substitution_array.push(value);
     }
@@ -137,16 +140,21 @@ pub fn read_template<'a>(
 pub fn read_template_definition<'a>(
     cursor: &mut Cursor<&'a [u8]>,
     chunk: Option<&'a EvtxChunk<'a>>,
+    ansi_codec: EncodingRef,
 ) -> Result<BinXMLTemplateDefinition<'a>> {
     let next_template_offset = try_read!(cursor, u32);
 
     let template_guid = try_read!(cursor, guid);
-    let data_size = try_read!(cursor, u32);
-
     // Data size includes the fragment header, element and end of file token;
     // except for the first 33 bytes of the template definition (above)
-    let _data = *cursor.get_ref();
-    let tokens = BinXmlDeserializer::read_binxml_fragment(cursor, chunk, Some(data_size), false)?;
+    let data_size = try_read!(cursor, u32);
+    let tokens = BinXmlDeserializer::read_binxml_fragment(
+        cursor,
+        chunk,
+        Some(data_size),
+        false,
+        ansi_codec,
+    )?;
 
     Ok(BinXMLTemplateDefinition {
         next_template_offset,
@@ -243,6 +251,7 @@ mod test {
     use crate::binxml::value_variant::BinXmlValue;
     use crate::ensure_env_logger_initialized;
     use crate::guid::Guid;
+    use encoding::all::WINDOWS_1252;
     use std::borrow::Cow;
     use std::io::{Cursor, Seek, SeekFrom};
 
@@ -487,7 +496,7 @@ mod test {
 
         let mut c = Cursor::new(from_start_of_chunk);
         c.seek(SeekFrom::Start(550)).unwrap();
-        let actual = read_template_definition(&mut c, None).unwrap();
+        let actual = read_template_definition(&mut c, None, WINDOWS_1252).unwrap();
 
         assert_eq!(actual, expected_at_550);
     }