+++ src/utils/binxml_utils.rs
@@ -1,6 +1,10 @@
 use crate::evtx_parser::ReadSeek;
 
+use crate::err::{self, Result};
+
 use byteorder::{LittleEndian, ReadBytesExt};
+
+use encoding::{decode, DecoderTrap, EncodingRef};
 use log::{error, trace};
 use std::char::decode_utf16;
 use std::io::{self, Error, ErrorKind};
@@ -54,6 +58,37 @@ pub fn read_utf16_by_size<T: ReadSeek>(stream: &mut T, size: u64) -> io::Result<
             }
             Some(s)
         }),
+    }
+}
+
+/// Reads an ansi encoded string from the given stream using `ansi_codec`.
+pub fn read_ansi_encoded_string<T: ReadSeek>(
+    stream: &mut T,
+    size: u64,
+    ansi_codec: EncodingRef,
+) -> Result<Option<String>> {
+    match size {
+        0 => Ok(None),
+        _ => {
+            let mut bytes = vec![0; size as usize];
+            stream.read_exact(&mut bytes)?;
+
+            let s = match decode(&bytes, DecoderTrap::Strict, ansi_codec).0 {
+                Ok(mut s) => {
+                    if let Some('\0') = s.chars().last() {
+                        s.pop();
+                    }
+                    s
+                }
+                Err(message) => Err(err::Error::FailedToDecodeANSIString {
+                    encoding: ansi_codec.name(),
+                    message: message.to_string(),
+                    offset: stream.tell()?,
+                })?,
+            };
+
+            Ok(Some(s))
+        }
     }
 }