+++ src/utils/mod.rs
@@ -3,7 +3,8 @@ mod hexdump;
 mod time;
 
 pub use self::binxml_utils::{
-    read_len_prefixed_utf16_string, read_null_terminated_utf16_string, read_utf16_by_size,
+    read_ansi_encoded_string, read_len_prefixed_utf16_string, read_null_terminated_utf16_string,
+    read_utf16_by_size,
 };
 pub use self::hexdump::{dump_cursor, print_hexdump};
 pub use self::time::{datetime_from_filetime, read_systemtime};