+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+Implemented Ansi codecs! 
+
+### Added
+- `--ansi-codec` to control the codec that will be used to decode ansi encoded strings inside the document.
+
+
 ## [0.3.0] - 2019-05-14
 
 This is a minor release due to the removal of `failure`.